+++ test/time_zone_info/data_config_test.exs
@@ -27,20 +27,80 @@ defmodule TimeZoneInfo.DataConfigTest do
                {:error, {:time_zones_not_found, ["Foo"]}}
     end
 
+    @tag :only
     test "returns time zones for configured link", %{data: data} do
       time_zones = ["Europe/Jersey"]
       assert {:ok, data} = DataConfig.update_time_zones(data, time_zones)
       assert Map.keys(data.time_zones) == ["Europe/London"]
       assert Map.keys(data.rules) == ["EU"]
+      assert Map.keys(data.links) == ["Europe/Jersey"]
+    end
 
-      assert Map.keys(data.links) == [
-               "Europe/Belfast",
-               "Europe/Guernsey",
-               "Europe/Isle_of_Man",
-               "Europe/Jersey",
-               "GB",
-               "GB-Eire"
+    test "returns configured time zones for africa", %{data: data} do
+      time_zones = ["Africa"]
+      assert {:ok, data} = DataConfig.update_time_zones(data, time_zones)
+
+      assert data.time_zones |> Map.keys() |> Enum.sort() == [
+               "Africa/Abidjan",
+               "Africa/Accra",
+               "Africa/Algiers",
+               "Africa/Bissau",
+               "Africa/Cairo",
+               "Africa/Casablanca",
+               "Africa/Ceuta",
+               "Africa/El_Aaiun",
+               "Africa/Johannesburg",
+               "Africa/Juba",
+               "Africa/Khartoum",
+               "Africa/Lagos",
+               "Africa/Maputo",
+               "Africa/Monrovia",
+               "Africa/Nairobi",
+               "Africa/Ndjamena",
+               "Africa/Sao_Tome",
+               "Africa/Tripoli",
+               "Africa/Tunis",
+               "Africa/Windhoek"
              ]
+
+      assert data.links |> Map.keys() |> Enum.sort() == [
+               "Africa/Addis_Ababa",
+               "Africa/Asmara",
+               "Africa/Asmera",
+               "Africa/Bamako",
+               "Africa/Bangui",
+               "Africa/Banjul",
+               "Africa/Blantyre",
+               "Africa/Brazzaville",
+               "Africa/Bujumbura",
+               "Africa/Conakry",
+               "Africa/Dakar",
+               "Africa/Dar_es_Salaam",
+               "Africa/Djibouti",
+               "Africa/Douala",
+               "Africa/Freetown",
+               "Africa/Gaborone",
+               "Africa/Harare",
+               "Africa/Kampala",
+               "Africa/Kigali",
+               "Africa/Kinshasa",
+               "Africa/Libreville",
+               "Africa/Lome",
+               "Africa/Luanda",
+               "Africa/Lubumbashi",
+               "Africa/Lusaka",
+               "Africa/Malabo",
+               "Africa/Maseru",
+               "Africa/Mbabane",
+               "Africa/Mogadishu",
+               "Africa/Niamey",
+               "Africa/Nouakchott",
+               "Africa/Ouagadougou",
+               "Africa/Porto-Novo",
+               "Africa/Timbuktu"
+             ]
+
+      assert Map.keys(data.rules) == ["EU"]
     end
 
     test "returns configured time zones", %{data: data} do
@@ -48,6 +108,7 @@ defmodule TimeZoneInfo.DataConfigTest do
       assert {:ok, data} = DataConfig.update_time_zones(data, time_zones)
 
       assert Map.keys(data.time_zones) == [
+               "Africa/Nairobi",
                "America/Indiana/Indianapolis",
                "America/Indiana/Knox",
                "America/Indiana/Marengo",
@@ -70,17 +131,9 @@ defmodule TimeZoneInfo.DataConfigTest do
       assert Map.keys(data.rules) == ["EU", "US"]
 
       assert Map.keys(data.links) == [
-               "America/Fort_Wayne",
-               "America/Indianapolis",
-               "America/Knox_IN",
-               "Europe/Belfast",
-               "Europe/Guernsey",
-               "Europe/Isle_of_Man",
-               "Europe/Jersey",
-               "GB",
-               "GB-Eire",
-               "US/East-Indiana",
-               "US/Indiana-Starke"
+               "Indian/Antananarivo",
+               "Indian/Comoro",
+               "Indian/Mayotte"
              ]
     end
   end