+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Refactoring `updater` to fix some unexpected behaviours when changing configs.
 - Updating docs structure.
 - Adding `backward` to the default `files`.
+- Refactoring `time_zones` configuration behaviour.
+  See [Time zone maniputlation](time_zone_info/config.html#time-zone-manipulation).
 
 ## 0.4.0 - 2020/04/25